refactor(LoggedInUserContext): change structure
